Move tests to subfolder
Now the tests folder contains pylint and simpleline_tests folders.
