+++ tests/simpleline_tests/containers_test.py
@@ -18,15 +18,14 @@
 #
 
 import unittest
-from unittest.mock import patch
 from io import StringIO
-
-from tests.widgets_test import BaseWidgets_TestCase
+from unittest.mock import patch
 
 from simpleline import App
 from simpleline.render.containers import WindowContainer, ListRowContainer, ListColumnContainer, KeyPattern
 from simpleline.render.screen import UIScreen, InputState
 from simpleline.render.widgets import TextWidget
+from tests.simpleline_tests.widgets_test import BaseWidgets_TestCase
 
 
 class Containers_TestCase(BaseWidgets_TestCase):
