+++ tests/simpleline_tests/glib_tests/screen_scheduler_glib_test.py
@@ -1,4 +1,4 @@
-# Event loop test classes for GLib implementation.
+# Screen scheduling GLib test classes.
 #
 # Copyright (C) 2017  Red Hat, Inc.
 #
@@ -17,16 +17,21 @@
 # Red Hat, Inc.
 #
 
-from tests.event_loop_test import ProcessEvents_TestCase
-from tests.glib_tests import GLibUtilityMixin
+
+from tests.simpleline_tests.glib_tests import GLibUtilityMixin
+from tests.simpleline_tests.screen_scheduler_test import ScreenScheduler_TestCase
 
 
-class GLibProcessEvents_TestCase(ProcessEvents_TestCase, GLibUtilityMixin):
-    """Run all the tests in ProcessEvents test case but with GLib event loop."""
+class GLibScrenScheduler_TestCase(ScreenScheduler_TestCase, GLibUtilityMixin):
+
+    def setUp(self):
+        super().setUp()
+        self.loop = None
+        self.timeout_error = False
 
     def tearDown(self):
         super().tearDown()
         self.teardown_glib()
 
-    def create_loop(self):
-        self.create_glib_loop()
+    def schedule_screen_and_run(self, screen):
+        self.schedule_screen_and_run_with_glib(screen)
