+++ tests/simpleline_tests/render_screen_test.py
@@ -25,7 +25,7 @@ from unittest import mock
 from simpleline import App
 from simpleline.render import RenderUnexpectedError
 from simpleline.render.screen import UIScreen, InputState
-from tests import UtilityMixin
+from tests.simpleline_tests import UtilityMixin
 
 
 def _fake_input(queue_instance, prompt, hidden):
